updated web.php, actorcontroller in theory working
